Move TopicFilter from network codec to json_rpc::methods
